+++ src/components/sensCalcDisplay/single/SensCalcDisplaySingle.cy.tsx
@@ -7,7 +7,7 @@ import SensCalcDisplaySingle from './SensCalcDisplaySingle';
 const THEME = [THEME_DARK, THEME_LIGHT];
 
 function mountingBasic(theTheme: any, show: boolean) {
-  cy.viewport(1500, 1500);
+  cy.viewport(2000, 1000);
   cy.mount(
     <ThemeProvider theme={theme(theTheme)}>
       <CssBaseline />