+++ src/services/axios/getProposalList/getProposalList.tsx
@@ -45,30 +45,32 @@ const getScienceCategory = (scienceCat: string) => {
 };
 
 const getPI = (investigators: InvestigatorBackend[]): string => {
-  const principalInvestigator = investigators.find(p => p.principal_investigator === true);
-  return `${principalInvestigator.given_name} ${principalInvestigator.family_name}`;
+  const principalInvestigator = investigators?.find(p => p.principal_investigator === true);
+  return `${principalInvestigator?.given_name} ${principalInvestigator?.family_name}`;
 };
 
 function mappingList(inRec: ProposalBackend[]): Proposal[] {
   const output = [];
   for (let i = 0; i < inRec.length; i++) {
     const rec: Proposal = {
-      id: inRec[i].prsl_id.toString(),
+      id: inRec[i].prsl_id?.toString(),
       status: inRec[i].status,
-      lastUpdated: new Date(inRec[i].metadata.last_modified_on).toDateString(),
-      lastUpdatedBy: inRec[i].metadata.last_modified_by,
-      createdOn: inRec[i].metadata.created_on,
-      createdBy: inRec[i].metadata.created_by,
-      version: inRec[i].metadata.version,
-      proposalType: Projects.find(p => p.mapping === inRec[i].info.proposal_type.main_type)?.id,
-      proposalSubType: inRec[i].info.proposal_type.sub_type
-        ? getSubType(inRec[i].info.proposal_type)
+      lastUpdated: inRec[i].metadata?.last_modified_on,
+      lastUpdatedBy: inRec[i].metadata?.last_modified_by,
+      createdOn: inRec[i].metadata?.created_on,
+      createdBy: inRec[i].metadata?.created_by,
+      version: inRec[i].metadata?.version,
+      proposalType: Projects.find(p => p.mapping === inRec[i].info?.proposal_type.main_type)?.id,
+      proposalSubType: inRec[i].info?.proposal_type.sub_type
+        ? getSubType(inRec[i].info?.proposal_type)
         : [],
-      scienceCategory: getScienceCategory(inRec[i].info.science_category),
-      title: inRec[i].info.title,
-      cycle: inRec[i].cycle,
-      team: getTeam(inRec[i].info.investigators),
-      pi: getPI(inRec[i].info.investigators)
+      scienceCategory: inRec[i].info?.science_category
+        ? getScienceCategory(inRec[i].info.science_category)
+        : null,
+      title: inRec[i].info?.title,
+      cycle: inRec[i]?.cycle,
+      team: inRec[i].info?.investigators ? getTeam(inRec[i].info.investigators) : [],
+      pi: inRec[i].info?.investigators ? getPI(inRec[i].info.investigators) : ''
       // telescope: 'N/A' // TODO is this still needed? -> what to map to? telescopes in observations?
     };
     output.push(rec);